Create SettingsDataStore.kt for managing session variables like DarkTheme.
