+++ app/src/main/java/com/codingwithmitch/food2forkcompose/presentation/MainActivity.kt
@@ -9,6 +9,7 @@ import androidx.compose.ui.viewinterop.viewModel
 import androidx.hilt.navigation.HiltViewModelFactory
 import androidx.navigation.NavType
 import androidx.navigation.compose.*
+import com.codingwithmitch.food2forkcompose.datastore.SettingsDataStore
 import com.codingwithmitch.food2forkcompose.presentation.navigation.Screen.RecipeDetail
 import com.codingwithmitch.food2forkcompose.presentation.navigation.Screen.RecipeList
 import com.codingwithmitch.food2forkcompose.presentation.ui.recipe.RecipeDetailScreen
@@ -27,6 +28,9 @@ class MainActivity : AppCompatActivity() {
 
   @Inject
   lateinit var connectivityManager: ConnectivityManager
+
+  @Inject
+  lateinit var settingsDataStore: SettingsDataStore
 
   override fun onStart() {
     super.onStart()
@@ -47,10 +51,10 @@ class MainActivity : AppCompatActivity() {
           val factory = HiltViewModelFactory(AmbientContext.current, navBackStackEntry)
           val viewModel: RecipeListViewModel = viewModel("RecipeListViewModel", factory)
           RecipeListScreen(
-            isDarkTheme = (application as BaseApplication).isDark.value,
+            isDarkTheme = settingsDataStore.isDark.value,
             isNetworkAvailable = connectivityManager.isNetworkAvailable.value,
             onNavigateToRecipeDetailScreen = navController::navigate,
-            onToggleTheme = (application as BaseApplication)::toggleTheme,
+            onToggleTheme = settingsDataStore::toggleTheme,
             viewModel = viewModel,
           )
         }
@@ -63,7 +67,7 @@ class MainActivity : AppCompatActivity() {
           val factory = HiltViewModelFactory(AmbientContext.current, navBackStackEntry)
           val viewModel: RecipeDetailViewModel = viewModel("RecipeDetailViewModel", factory)
           RecipeDetailScreen(
-            isDarkTheme = (application as BaseApplication).isDark.value,
+            isDarkTheme = settingsDataStore.isDark.value,
             isNetworkAvailable = connectivityManager.isNetworkAvailable.value,
             recipeId = navBackStackEntry.arguments?.getInt("recipeId"),
             viewModel = viewModel,
